LayoutTests/imported/w3c/web-platform-tests/html/webappapis/animation-frames/idlharness-expected.txt

CONSOLE MESSAGE: line 309: callback not yet supported
idlharness test

This test validates the WebIDL included in the Timing control for script-based animations specification.

partial interface Window {
  long requestAnimationFrame(FrameRequestCallback callback);
  void cancelAnimationFrame(long handle);
};

callback FrameRequestCallback = void (DOMHighResTimeStamp time);

FAIL Window interface: operation requestAnimationFrame(FrameRequestCallback) assert_own_property: interface prototype object missing non-static operation expected property "requestAnimationFrame" missing
PASS Unscopable handled correctly for requestAnimationFrame(FrameRequestCallback) on Window 
FAIL Window interface: operation cancelAnimationFrame(long) assert_own_property: interface prototype object missing non-static operation expected property "cancelAnimationFrame" missing
PASS Unscopable handled correctly for cancelAnimationFrame(long) on Window 
FAIL Window interface: window must inherit property "requestAnimationFrame(FrameRequestCallback)" with the proper type assert_inherits: property "requestAnimationFrame" found on object expected in prototype chain
FAIL Window interface: calling requestAnimationFrame(FrameRequestCallback) on window with too few arguments must throw TypeError assert_inherits: property "requestAnimationFrame" found on object expected in prototype chain
FAIL Window interface: window must inherit property "cancelAnimationFrame(long)" with the proper type assert_inherits: property "cancelAnimationFrame" found on object expected in prototype chain
FAIL Window interface: calling cancelAnimationFrame(long) on window with too few arguments must throw TypeError assert_inherits: property "cancelAnimationFrame" found on object expected in prototype chain
